% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{add_params}
\alias{add_params}
\title{Add parameters}
\usage{
add_params(corsim, paramvec)
}
\arguments{
\item{corsim}{a list with the correlation values (output of \code{\link[=get_cor]{get_cor()}}).}

\item{paramvec}{A vector of parameters (must be a 1-row dataframe)}
}
\value{
The list \code{corsim} with added columns in each dataframe of the list.
The added columns are named like the columns of \code{paramvec} and contain repeated values
of \code{paramvec}
}
\description{
Adds parameters value to simulation results dataframe.
}
