% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{combine_cor}
\alias{combine_cor}
\title{Combine correlations}
\usage{
combine_cor(a, b)
}
\arguments{
\item{a}{First list formatted by \code{\link[=get_cor]{get_cor()}}}

\item{b}{Second list formatted by \code{\link[=get_cor]{get_cor()}}}
}
\value{
A list as formatted by \code{\link[=get_cor]{get_cor()}} (see this function's
"Return" section).
}
\description{
Combine lists formatted by \code{\link[=get_cor]{get_cor()}}.
}
