% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compas2d.R
\name{compas2d}
\alias{compas2d}
\title{Simulate species interactions}
\usage{
compas2d(
  nconsumer = 40,
  nresource = 100,
  le_grad = 100,
  ratio_grad = 0.8,
  consumer_ab = NULL,
  resource_ab = NULL,
  min_ab = 1,
  max_ab = 100,
  ninter = 100,
  delta = 1,
  mean_tol = 2,
  sd_tol = 10,
  buffer = 1,
  col_prefix = "c",
  row_prefix = "r",
  rowvar_prefix = "tr",
  remove_zeroes = FALSE,
  return_intermediate = FALSE
)
}
\arguments{
\item{nconsumer}{Number of consumer species}

\item{nresource}{Number of resource species}

\item{le_grad}{gradient length for traits. The values are comprised between
0 and \code{le_grad}.}

\item{ratio_grad}{length of the second gradient as a ratio of the first one.}

\item{consumer_ab}{Optional vector of abundances for consumers.
If missing, abundances are taken from a uniform law between \code{min_ab} and \code{max_ab}.}

\item{resource_ab}{Optional vector of abundances for resource species.
If missing, abundances are taken from a uniform law between \code{min_ab} and \code{max_ab}.}

\item{min_ab}{minimal abundance value for the uniform law
(if \code{consumer_ab} and/or \code{resource_ab} are absent)}

\item{max_ab}{maximal abundance value for the uniform law
(if \code{consumer_ab} and/or \code{resource_ab} are absent)}

\item{ninter}{Total number of observations in the matrix}

\item{delta}{Exponent between 0 and 1 to give more (1) or less (0)
weight to trait matching relatively to abundance.}

\item{mean_tol}{Mean value for the distribution of the normal law
in which consumers niche standard deviation are drawn (for each trait).}

\item{sd_tol}{Standard deviation for the distribution of the normal law
in which consumers niche standard deviation are drawn (for each trait).}

\item{buffer}{buffer to allow consumer species optima to fall outside the
gradient by 0 - \code{buffer} ; \code{le_grad} + \code{buffer}.}

\item{col_prefix}{Prefix for the column names of the matrix (which will be
prefix + index).
Defaults to "c" for "consumers".}

\item{row_prefix}{Prefix for the row names of the matrix (which will be
prefix + index).
Defaults to "r" for "resources".}

\item{rowvar_prefix}{Prefix for the column names of the row variable
(which will be prefix + 1 or 2).
Defaults to "tr" for "resource traits".}

\item{remove_zeroes}{If there are unobserved species,
should we keep them in the final matrix?}

\item{return_intermediate}{Return intermediate results? If yes, will return the
probability of interaction based on matching only, on abundances only and
the predicted mix of matching and abundances (before sampling).}
}
\value{
An object of class compas, which is a list with 3 components
(4 if \code{return_intermediate}):
\itemize{
\item \code{comm}: the interaction matrix of dimension \code{nresource} x \code{nconsumer}.
Rows and columns are named like species (see \code{row_prefix} and \code{col_prefix}).
\item \code{rowvar}: resource traits values. It is a \code{nresource} x 2 matrix.
Rows are named like resource species (see \code{row_prefix}) and
column are named like resource traits (see \code{rowvar_prefix}).
\item \code{colvar}: consumer traits values. It is a \code{nconsumer} x 2 x 2 array.
Rows are named like consumer species (see \code{col_prefix}).
Column names in each array dimension are \code{mean} and \code{sd}
(for the mean and standard deviation of the species niche, respectively).
\item if \code{return_intermediate}, a fourth component named intermediate is returned.
The first element is named \code{p_matching} and contains the matrix of
probabilities of interactions based only on matching.
The second \code{ab_neutral} contains the matrix of count of interactions as
predicted by species abundances.
The third \code{p_mix} contains the matrix of probabilities of
interactions taking into account the mix of matching and abundances.
}
}
\description{
Simulate species interactions based on the matching between 2 traits.
This is based on a code that originally simulated species abundances
in different sites based on their environmental niches.
}
\details{
This function models species interactions as arising from the matching between their traits.
Traits are drawn from a uniform law with parameters
determined by \code{le_grad} and \code{ratio_grad} (for the first and second trait).
The matching is based on a multivariate normal distribution with means
corresponding to the difference between trait values, and variance determined by
the variance of consumers' niches.
It is then mixed to the abundances (using the parameter \code{delta}) and sampled to
obtain \code{ninter}interactions.
}
