% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{filter_matrix}
\alias{filter_matrix}
\title{Filter matrix}
\usage{
filter_matrix(mat, thr)
}
\arguments{
\item{mat}{The incidence matrix.}

\item{thr}{The minimal number of different interacting partners a species
must have (a species must have \verb{>= thr} different interacting partners).
It is not weighted (for instance, if \code{thr = 2} then a species interacting 100 times
with a unique other partner will be removed.)}
}
\value{
The interaction matrix but where the rows and/or columns with species that do not interact
above the threshold removed.
}
\description{
Filter incidence matrix to keep only species that interact
with more partners that a given threshold.
}
