% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_best_model}
\alias{get_best_model}
\title{Get best model}
\usage{
get_best_model(lmsimple, lm2, method = c("LRT", "AIC"), alpha = 0.05)
}
\arguments{
\item{lmsimple}{First model (if using \code{method = "LRT"},
must be nested in \code{lm2})}

\item{lm2}{Second model}

\item{method}{The method to use to compare models:
likelihood ratio test (\code{method = "LRT"}) for nested linear
models comparison or AIC (\code{method = "AIC"})}

\item{alpha}{Significance threshold to consider when using
\code{method = "LRT"}}
}
\value{
The best model
}
\description{
Compare 2 linear models and returns the best
}
