% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{get_cor}
\alias{get_cor}
\title{Get correlations}
\usage{
get_cor(niches)
}
\arguments{
\item{niches}{An object containing the fundamental, realized and multivariate niches
(output of \code{\link[=get_niches]{get_niches()}})}
}
\value{
A list of lists. The first element (named \code{fundamental}) contains the correlations with the
fundamental niche values and the second element (named \code{realized})
contains the correlations with the realized niche values.
Each sublist is a list of 2 dataframes:
\itemize{
\item \code{mean_cor}: correlations of niches optima with the CA coordinates.
\item \code{sd_cor}: correlations of niches breadths with reciprocal scaling standard deviations.
}

These dataframes have the following columns:
\itemize{
\item \code{axis}: correlations measured on which multivariate axis?
\item \code{type}: the type of individuals for which the measure is (rows or columns?).
Column or row codes are chosen with \code{rowname} and \code{colname}
from the parameter \code{niches}.
\item \code{cor}: the correlation value
}
}
\description{
Measure correlations between niche obtained from CA or reciprocal scaling
and the fundamental and realized niches.
}
