% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{get_niches}
\alias{get_niches}
\title{Get niches}
\usage{
get_niches(
  rec,
  comm,
  consumer_niche,
  resource_traits,
  rowname = "row",
  colname = "col"
)
}
\arguments{
\item{rec}{A dataframe (expected to be the output of \code{\link[=reciprocal.coa]{reciprocal.coa()}}).
The first 2 columns must contain the reciprocal scaling scores (there can be more axes, but only
the first 2 will be used.) The last 3 columns are \code{Row}, \code{Col} and \code{Weight}.}

\item{comm}{The observed community matrix. It is used to get the
resource species niche width.}

\item{consumer_niche}{A nx2x2 array with n rows (n species),
2 columns corresponding to species niche optima and standard deviation
(corresponding to their multivariate normal niche) and 2 traits stored
in the dimension.
There is 1 row per species.}

\item{resource_traits}{A mx2 matrix. Each row corresponds to a
species and each column to a trait value.}

\item{rowname}{How to name the first element of the list
(resources/rows) (e.g. could be "plants")}

\item{colname}{How to name the second element of the list
(resources/rows) (e.g. could be "birds")}
}
\value{
A list of lists of lists. The first element (named \code{rowname}) contains the niches for the
resources and the second element (named \code{colname}) contains the niches for the consumers.
Each element contains 2 lists:
\itemize{
\item \code{mean}: niche optima
\item \code{sd}: niche breadths
}

For each of these sublists, there are 3 matrices elements named:
\itemize{
\item \code{fundamental}: the fundamental niche based on species traits
\item \code{realized}: the realized niche based on interacting species traits
\item \code{mvar}: niche parameters corresponding to the mean and standard deviations
given by reciprocal scaling.
}

These matrices describe the value for each species (in rows)
along each trait/axis (in column).
}
\description{
Returns the fundamental and realized niches and the niches returned
by reciprocal scaling.
}
\details{
The fundamental niche optimum of a consumer is the mean of its niche
(from the \code{consumer_niche} table).
For a resource, it corresponds to the traits from the \code{resource_traits} table.
The fundamental niche breadths are computed for consumers only,
and corresponds to the breadths of their niche (from the \code{consumer_niche} table).

The realized niche optima are computed as the weighted means of the traits
of species' interacting partners: therefore, the realized niche
of a consumer is the weighted mean of the traits of the resource species
it interacts with, and reciprocally.
The realized niche breadths are computed as the weighted standard deviations
of the traits of a species' interacting partners. Therefore, the realized
niche breadth of a consumer is the weighted standard deviation of the
traits of the resource species it interacts with, and reciprocally.

Niche optima computed with multivariate methods correspond to the reciprocal
scaling mean and niche breadths correspond to the reciprocal scaling
standard deviation.
}
