% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_corcircle}
\alias{plot_corcircle}
\title{Plot correlation circle}
\usage{
plot_corcircle(
  cor,
  label = rownames(cor),
  col = "black",
  lty = "solid",
  cor2 = NULL,
  label2 = rownames(cor2),
  col2 = "black",
  lty2 = "solid",
  xax = 1,
  yax = 2,
  xlab = NULL,
  ylab = NULL,
  eig = NULL,
  col_bg = "grey30",
  mar = 0.01,
  labsize = 3.88,
  lim = 1,
  clip = "off"
)
}
\arguments{
\item{cor}{dataframe containing the coordinates of the variables for each axis in columns}

\item{label}{label to display on the arrows for the
first set of variables}

\item{col}{color for the first set of arrows}

\item{lty}{linetype for the first set of variables}

\item{cor2}{optional second set of variables (for double constrained analyses)}

\item{label2}{label to display on the arrows for the
second set of variables}

\item{col2}{color for the second set of arrows}

\item{lty2}{linetype for the second set of variables}

\item{xax}{index of the column of the matrix to use for the x-axis}

\item{yax}{index of the column of the matrix to use for the y-axis}

\item{xlab}{Custom x-label}

\item{ylab}{Custom y-label}

\item{eig}{Eigenvalues vector (to display percentage of variation explained by each axis)}

\item{col_bg}{color to use for background elements
(circle and y = 0 and x = 0 lines)}

\item{mar}{margin between arrow tips and label}

\item{labsize}{Size for arrow labels.}

\item{lim}{limits for the x and y axes:
a unique number that gives the distance between the plot
limit and zero (so that the circle is centered)}

\item{clip}{constrain the labels to be inside the plot? If yes, change to "on".}
}
\value{
A ggplot with the correlation circle where one or two sets of variables
are represented as arrows inside the circle of radius 1.
}
\description{
Plot a correlation circle
}
