% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprocal_ade4.R
\name{reciprocal.coa}
\alias{reciprocal.coa}
\title{Reciprocal scaling for CA}
\usage{
reciprocal.coa(x)
}
\arguments{
\item{x}{output of \code{dudi.coa}}
}
\value{
Results of the reciprocal scaling.
Each row corresponds to a correspondence in the original table (a nonzero occurrence).

The first columns (\code{Scorexx}) give coordinates in the different dimensions.
\code{Scorexx} gives the coordinates of the correspondence in the multivariate
space (as given by canonical correlation analysis).

\code{Row} and \code{Col} give the row and column this correspondence belongs to.

\code{Weight} gives the correspondence weight (the count of the original cell divided
by the sum of the table).
}
\description{
Performs reciprocal scaling from corrrespondence analysis
}
\references{
See the original publication: Thioulouse, J., & Chessel, D. (1992).
A Method for Reciprocal Scaling of Species Tolerance and Sample Diversity.
Ecology, 73(2), 670–680. \doi{doi:10.2307/1940773}
}
