% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSnetwork-package.R
\docType{package}
\name{RSnetwork-package}
\alias{RSnetwork}
\alias{RSnetwork-package}
\title{RSnetwork: Ecological Networks Analysis with Reciprocal Scaling}
\description{
Analyze ecological networks and possibly species-environment tables using various correspondence analyses methods, and also quantify niches widths with reciprocal scaling.
}
\author{
\strong{Maintainer}: Lisa Nicvert \email{lisa.nicvert@univ-lyon1.fr} (\href{https://orcid.org/0009-0006-5763-0865}{ORCID})

}
\keyword{internal}
