% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_eig}
\alias{plot_eig}
\title{Plot eigenvalues}
\usage{
plot_eig(eigenvalues, showrank = FALSE)
}
\arguments{
\item{eigenvalues}{The eigenvalues}

\item{showrank}{Should the x-axis display the rank of the eigenvalues?}
}
\value{
A ggplot with the eigenvalues displayed as a barplot.
}
\description{
Barplot of eigenvalues
}
